Update BroadcastExecutor.java
A pixelmon server owner complained about broadcast links not functioning, I took upon myself to find a way for links to be recognized and applied to the text. If the message contained a link I cut it into three pars the link, everything before, and everything after. If there was an easier method I wouldn't mind hearing, but this was all I could manage. The same owner found that chat links are also non-functional and I'm sure this fix can be applied to them. Should you need to contact this owner, http://kpixelmon.enjin.com/. He has no clue how to submit a bug himself.